requirements.txt
Flask>=2.3
Flask-JWT-Extended>=4.6
python-dotenv>=1.0
psycopg2-binary>=2.9
Flask-Migrate>=4.0
Flask-CORS>=3.0
gunicorn>=21.2
bcrypt>=4.1
-r yolov5/requirements.txt
